[Multi Inheritance Implementation] Now Working...
IT WORKS! We are now literally learning something from our discovery. Now we have to connect other window.
